Configure a set of model fits from a specification of different combinations

Configure a set of model fits from a specification of different combinations

Usage

configure_models(
  model_formula,
  dataset,
  model_function,
  data_subset = data_subset_provider(default = TRUE),
  formula_update = formula_provider(default = . ~ .)
)

Arguments

model_formula

the model formula or formulae you want to fit as a formula_provider(...).

dataset

the dataset imputed or bootstrapped if need be as a bootstrap_provider(...) or imputation_provider(...)

model_function

the model functions to fit as a model_function_provider(...)

data_subset

the data subsets to fit as a data_subset_provider(...)

formula_update

and model formula updates to apply before fitting as a formula_provider(...).

Value

a configuration dataframe and data provider to run

Examples

form = ~ bmi + age + chl

fp = formula_provider(
  `Univariate` = modelfitter::univariate_from_multivariate(form),
  `Adj Model 1` = form,
  `Adj Model 2` = update(form, . ~ . - chl),
)
#> Error in formula_provider(Univariate = modelfitter::univariate_from_multivariate(form),     `Adj Model 1` = form, `Adj Model 2` = update(form, . ~ . -         chl), ): could not find function "formula_provider"

mfp = model_function_provider(
   `Logistic` = modelfitter::logistic_regression,
   `Poisson` = modelfitter::quasi_poisson
)
#> Error in model_function_provider(Logistic = modelfitter::logistic_regression,     Poisson = modelfitter::quasi_poisson): could not find function "model_function_provider"

ip = imputation_provider(
    mice::nhanes2 %>% dplyr::mutate(death=TRUE), 10)
#> using cached item: ~/.cache/modelfitter/cached-d485da4f27080dab75135297c7bda544-a41a192ca85a1a7696bcd194f044c399.rda

fup = formula_provider(
  hypertension = hyp ~ .,
  death = death ~ .
)
#> Error in formula_provider(hypertension = hyp ~ ., death = death ~ .): could not find function "formula_provider"

dsp = data_subset_provider(
  all = TRUE,
  hypertensives = hyp == "yes",
  nonhypertensives = hyp == "no" 
)

tmp = configure_models(fp, ip, mfp, dsp, fup) %>% dplyr::glimpse()
#> Error in eval(expr, envir, enclos): object 'fp' not found

# the number of models to fit will be this:
sum(tmp$n_boots)
#> Error in eval(expr, envir, enclos): object 'tmp' not found