Generate a table for a single model output

The modelfitter configuration, execution, summary, format pipeline is for the complex use cases with multiple models. This is the simple version when we have a single fitted model and we want to print it in a table. This is more or less what gtsummary does.

Usage

format_model(model_name, fit, statistic, ...)

Arguments

model_name

a title for the model.

fit

the model fit.

statistic

the type of stastic this model outputs (e.g. OR, HR, RR). This mostly defines the label in the table.

...

Arguments passed on to format_summary

summfit

A set of fitted, and summarised configured models

global.p

present the global p value (anova III) rather than the line by line values.

inv_link

the inverse of the link function employed in the models. This is almost always the inverse exp(...) unless we are dealing with a linear model.

col_header

a glue spec using columns from the summfit data table to label the model columns. model_name and n_obs_summary should be defined as a minimum. Other bits of metadata will be present if the table has been configured using configure_models(...) including model_type_name, data_subset_name, model_base_name, model_update_name, n_boots.

row_design

a glue spec for presenting the statistic. valid columns are reference - the referent status, group.type, beta.lower, beta.median, beta.upper, value.lower, value.median, value.upper, p.value.mixture, global.p.mixture, global.p.method. The helper functions format_ratio(x) and format_ci(med,low,hi,ref) may be useful in this glue string

p_format

a function (or lambda) converting a number into a p-value string

font_size

(optional) the font size for the table in points

font

(optional) the font family for the table (which will be matched to closest on your system)

footer_text

any text that needs to be added at the end of the table, setting this to FALSE dsables the whole footer (as does options("tableone.hide_footer"=TRUE)).

Value

a huxtable formatted model summary table

Examples

coxfit = cox_model(survival::flchain, survival::Surv(futime, death) ~ sex * age)
format_model("FLChain", coxfit, "HR")
#> Error in map2(.x, vec_index(.x), .f, ...): ℹ In index: 1.
#> Caused by error in `.f()`:
#> ! argument "label_fn" is missing, with no default