Format a summary of multiple fits into a table.

Format a summary of multiple fits into a table.

Usage

format_summary(
  summfit,
  ...,
  statistic = "OR",
  global.p = getOption("modelfitter.global_p_values", TRUE),
  inv_link = exp,
  col_header = "{model_name} (N={sprintf('%d',n_obs_summary)})",
  row_design = "{format_ci(value.median,value.lower,value.upper,reference)}",
  p_format = NULL,
  font_size = getOption("modelfitter.font_size", 8),
  font = getOption("modelfitter.font", "Arial"),
  footer_text = NULL
)

Arguments

summfit

A set of fitted, and summarised configured models

...

Arguments passed on to model_labels

model

a model or list of models.

label_fn

a function that allows a predictor label to be renamed. This should expect a vector and return a vector of the same length. Levels will be terms in the model function and may be column names, or combinations thereof

subgroup_label_fn

a function that allows a subgroup label to be renamed. This should expect a vector and return a vector of the same length. The input to this function will be either a factor level name or a combination of them or whatever else the model decides to name it's coefficients.

statistic

what model output is this table presenting? Is it an OR, a RR or a HR? or something else?

global.p

present the global p value (anova III) rather than the line by line values.

inv_link

the inverse of the link function employed in the models. This is almost always the inverse exp(...) unless we are dealing with a linear model.

col_header

a glue spec using columns from the summfit data table to label the model columns. model_name and n_obs_summary should be defined as a minimum. Other bits of metadata will be present if the table has been configured using configure_models(...) including model_type_name, data_subset_name, model_base_name, model_update_name, n_boots.

row_design

a glue spec for presenting the statistic. valid columns are reference - the referent status, group.type, beta.lower, beta.median, beta.upper, value.lower, value.median, value.upper, p.value.mixture, global.p.mixture, global.p.method. The helper functions format_ratio(x) and format_ci(med,low,hi,ref) may be useful in this glue string

p_format

a function (or lambda) converting a number into a p-value string

font_size

(optional) the font size for the table in points

font

(optional) the font family for the table (which will be matched to closest on your system)

footer_text

any text that needs to be added at the end of the table, setting this to FALSE dsables the whole footer (as does options("tableone.hide_footer"=TRUE)).

Value

a huxtable tabular output of the model(s)

Examples

cfg  = configure_models(
   formula_provider(
      "<F" = I(color < "F") ~ cut + carat + clarity + price,
      "<H" = I(color < "H") ~ cut + carat + clarity + price
   ),
   bootstrap_provider(ggplot2::diamonds, max_n = 100),
   model_function_provider(
     "Log reg" = modelfitter::logistic_regression,
     "Poisson" = modelfitter::quasi_poisson
   )
)
#> Error in formula_provider(`<F` = I(color < "F") ~ cut + carat + clarity +     price, `<H` = I(color < "H") ~ cut + carat + clarity + price): could not find function "formula_provider"

exectn = cfg %>% execute_configuration(cache = TRUE)
#> Error in eval(expr, envir, enclos): object 'cfg' not found
summfit = exectn %>% summarise_fits()
#> Error in eval(expr, envir, enclos): object 'exectn' not found
hux = summfit %>% format_summary()
#> Error in eval(expr, envir, enclos): object 'summfit' not found
hux
#> Error in eval(expr, envir, enclos): object 'hux' not found