Predict the potential coefficient names for a model output

Usage

format_summary_rows(rawData, ..., label_fn = ~.x)

Arguments

rawData

a modelfit or a dataframe(s)

...

the columns that we are using as predictors, as a list of formulae (rhs), a tidyselect call, a dplyr::vars() specification or a list of characters

label_fn

a function that converts column names into readable labels

Value

a dataframe with the likely names of the coefficients in a model which can be joined to a model coefficients to

Examples

boot = run_model(
  ggplot2::diamonds %>% dplyr::mutate(cut = factor(cut, ordered=FALSE)), 
  price ~ ., 
  stats::lm
)
#> Error in lifecycle::deprecate_soft("0.0.0.9011"): argument "what" is missing, with no default
tmp = format_summary_rows(boot)
#> Error in lifecycle::deprecate_soft("0.0.0.9011"): argument "what" is missing, with no default

# logistic regression (OR)
diamonds2 = ggplot2::diamonds %>% dplyr::mutate(
   is_coloured = color <= "F",
   dplyr::across(dplyr::where(is.ordered), ~ factor(.x,ordered=FALSE))
) %>% dplyr::select(-color)

format_summary_rows(diamonds2, is_coloured ~ cut + carat + clarity*price)
#> Error in lifecycle::deprecate_soft("0.0.0.9011"): argument "what" is missing, with no default
model = logistic_regression(diamonds2, is_coloured ~ cut + carat + clarity * price)
model = logistic_regression(diamonds2, is_coloured ~ cut + I(cut=="D") + clarity*price)

stats::coef(model)
#>       (Intercept)           cutGood      cutVery Good        cutPremium 
#>     -1.899078e-01      2.644440e-01      2.655568e-01      9.672590e-02 
#>          cutIdeal I(cut == "D")TRUE        claritySI2        claritySI1 
#>      2.243786e-01                NA      5.508182e-01      4.440135e-01 
#>        clarityVS2        clarityVS1       clarityVVS2       clarityVVS1 
#>      5.261397e-01     -1.463036e-01      6.243781e-02     -2.575117e-01 
#>         clarityIF             price  claritySI2:price  claritySI1:price 
#>     -8.898853e-01     -1.063789e-04      5.043942e-06     -1.832336e-07 
#>  clarityVS2:price  clarityVS1:price clarityVVS2:price clarityVVS1:price 
#>     -1.166712e-05      4.481136e-05      7.727660e-05      1.185510e-04 
#>   clarityIF:price 
#>      1.799989e-04